+++ lib/flutter_reactive_n.dart
@@ -1,0 +1,6 @@
+import 'package:flutter_reactive/flutter_reactive.dart';
+
+/// Nullable Reactive
+class ReactiveN<T> extends Reactive<T?> {
+  ReactiveN([super._value, super._strict]);
+}