+++ test/flutter_reactive_test.dart
@@ -1,3 +1,4 @@
+// ignore_for_file: public_member_api_docs, sort_constructors_first
 import 'package:flutter/material.dart';
 import 'package:flutter_reactive/flutter_reactive.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -33,7 +34,7 @@ void main() {
   test("Combine reactives", () {
     final active = true.reactive();
     final count = 0.reactive();
-    final message = ''.reactive();
+    final message = ReactiveN<String>();
     final status = Reactive.combine3(
       active,
       count,
@@ -50,6 +51,29 @@ void main() {
     message.value = 'Hello';
     active.toggle();
   });
+
+  test("User model", () {
+    final user = ReactiveN<UserModel>();
+    user.listen((value) {
+      print("$value");
+    });
+
+    user.value = UserModel(name: "Max", age: 22);
+
+    user.mutate((u) {
+      u?.name = "oedo";
+    });
+
+  });
+}
+
+class UserModel {
+  String name;
+  int age;
+  UserModel({required this.name, required this.age});
+
+  @override
+  String toString() => 'UserModel(name: $name, age: $age)';
 }
 
 class _Test extends StatefulWidget {