+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 0.0.3
+
+- Added ReactiveN for nullable types
+- New debounce, mutate and computed methods
+- Updated README with new features
+
+## 0.0.2
+
+- Minor fixes and improvements
+
 ## 0.0.1
 
 - Initial release